select/unselect favorite station

+++ app/src/main/java/io/github/vladimirmi/localradio/data/source/CacheSource.java
@@ -7,6 +7,7 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
+import java.util.List;
 import java.util.Locale;
 
 import okhttp3.HttpUrl;
@@ -17,15 +18,21 @@ import okhttp3.internal.http.RealResponseBody;
 import okhttp3.internal.io.FileSystem;
 import okio.BufferedSource;
 import okio.Okio;
+import timber.log.Timber;
 
 /**
  * Created by Vladimir Mikhalev 12.04.2018.
  */
 public class CacheSource implements Interceptor {
 
-    private static final String PREFIX = "query_cache";
+
+    private static final String SEARCH_PREFIX = "search_cache";
+    private static final String URL_PREFIX = "url_cache";
     private static final String SUFFIX = ".json";
     private static final SimpleDateFormat dateFormat = new SimpleDateFormat("ddMMyyyy", Locale.ENGLISH);
+
+    private static final long SEARCH_CACHE_SIZE = 10 * 1024 * 1024; //10Mb
+    private static final int URL_CACHE_SIZE = 100; //files
 
     private final File cacheDir;
 
@@ -36,12 +43,15 @@ public class CacheSource implements Interceptor {
     @Override
     public Response intercept(Chain chain) throws IOException {
 
-        File cacheFile = createLocationQueryCache(chain.request().url());
+        File cacheFile = createLocationSearchCache(chain.request().url());
         if (cacheFile == null) {
-            cacheFile = createCoordinatesQueryCache(chain.request().url());
+            cacheFile = createCoordinatesSearchCache(chain.request().url());
         }
         if (cacheFile == null) {
-            cacheFile = createIpQueryCache(chain.request().url());
+            cacheFile = createIpSearchCache(chain.request().url());
+        }
+        if (cacheFile == null) {
+            cacheFile = createUrlQueryCache(chain.request().url());
         }
         if (cacheFile == null) {
             return chain.proceed(chain.request());
@@ -72,43 +82,73 @@ public class CacheSource implements Interceptor {
     }
 
     public void cleanCache() {
-        File[] files = cacheDir.listFiles((dir, name) -> name.startsWith(PREFIX));
-        for (File file : files) {
-            file.delete();
+        File[] files = cacheDir.listFiles((dir, name) -> name.startsWith(SEARCH_PREFIX));
+        long cacheSize = 0;
+        for (int i = files.length - 1; i >= 0; i--) {
+            File file = files[i];
+            cacheSize += file.length();
+            if (cacheSize > SEARCH_CACHE_SIZE) {
+                file.delete();
+                Timber.e("cleanCache: " + file.getName());
+            }
+        }
+
+        files = cacheDir.listFiles((dir, name) -> name.startsWith(URL_PREFIX));
+        cacheSize = 0;
+        for (int i = files.length - 1; i >= 0; i--) {
+            cacheSize++;
+            File file = files[i];
+            if (cacheSize > URL_CACHE_SIZE) {
+                file.delete();
+                Timber.e("cleanCache: " + file.getName());
+            }
         }
     }
 
-    private File createLocationQueryCache(HttpUrl url) {
+    private File createLocationSearchCache(HttpUrl url) {
         String country = url.queryParameter("country");
         String city = url.queryParameter("city");
 
         if (country == null || city == null) {
             return null;
         }
-        String fileName = String.format("%s_%s_%s_%s_%s", PREFIX, country,
-                city, dateFormat.format(new Date()), SUFFIX);
+        String fileName = String.format("%s_%s_%s_%s_%s", SEARCH_PREFIX, dateFormat.format(new Date()),
+                country, city, SUFFIX);
 
         return new File(cacheDir, fileName);
     }
 
-    private File createCoordinatesQueryCache(HttpUrl url) {
+    private File createCoordinatesSearchCache(HttpUrl url) {
         String latitude = url.queryParameter("latitude");
         String longitude = url.queryParameter("longitude");
 
         if (latitude == null || longitude == null) {
             return null;
         }
-        String fileName = String.format("%s_%s_%s_%s_%s", PREFIX, latitude,
-                longitude, dateFormat.format(new Date()), SUFFIX);
+        String fileName = String.format("%s_%s_%s_%s_%s", SEARCH_PREFIX, dateFormat.format(new Date()),
+                latitude, longitude, SUFFIX);
 
         return new File(cacheDir, fileName);
     }
 
-    private File createIpQueryCache(HttpUrl url) {
+    private File createIpSearchCache(HttpUrl url) {
         String ip = url.queryParameter("ip");
 
         if (ip == null) return null;
-        String fileName = String.format("%s_%s_%s_%s", PREFIX, ip, dateFormat.format(new Date()), SUFFIX);
+        String fileName = String.format("%s_%s_%s_%s", SEARCH_PREFIX, dateFormat.format(new Date()), ip, SUFFIX);
+
+        return new File(cacheDir, fileName);
+    }
+
+    private File createUrlQueryCache(HttpUrl url) {
+        List<String> pathSegments = url.pathSegments();
+        String id = url.queryParameter("station_id");
+
+        if (id == null || !pathSegments.get(pathSegments.size() - 1).equals("uberstationurl.php")) {
+            return null;
+        }
+
+        String fileName = String.format("%s_%s_%s_%s", URL_PREFIX, dateFormat.format(new Date()), id, SUFFIX);
 
         return new File(cacheDir, fileName);
     }
